requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
pandas>=2.2.2
python-dotenv>=1.0.1
playwright>=1.42.0
